+++ src/main/java/ee/carlrobert/codegpt/settings/service/azure/AzureSettingsState.java
@@ -1,24 +1,14 @@
 package ee.carlrobert.codegpt.settings.service.azure;
 
-import com.intellij.util.xmlb.annotations.Transient;
 import java.util.Objects;
 
 public class AzureSettingsState {
 
-  private static final String BASE_PATH = "/openai/deployments/%s/chat/completions?api-version=%s";
-
   private String resourceName = "";
   private String deploymentId = "";
   private String apiVersion = "";
-  private String baseHost = "https://%s.openai.azure.com";
-  private String path = BASE_PATH;
   private boolean useAzureApiKeyAuthentication = true;
   private boolean useAzureActiveDirectoryAuthentication;
-
-  @Transient
-  public boolean isUsingCustomPath() {
-    return !BASE_PATH.equals(path);
-  }
 
   public String getResourceName() {
     return resourceName;
@@ -42,22 +32,6 @@ public class AzureSettingsState {
 
   public void setApiVersion(String apiVersion) {
     this.apiVersion = apiVersion;
-  }
-
-  public String getBaseHost() {
-    return baseHost;
-  }
-
-  public void setBaseHost(String baseHost) {
-    this.baseHost = baseHost;
-  }
-
-  public String getPath() {
-    return path;
-  }
-
-  public void setPath(String path) {
-    this.path = path;
   }
 
   public boolean isUseAzureApiKeyAuthentication() {
@@ -90,14 +64,12 @@ public class AzureSettingsState {
         && useAzureActiveDirectoryAuthentication == that.useAzureActiveDirectoryAuthentication
         && Objects.equals(resourceName, that.resourceName)
         && Objects.equals(deploymentId, that.deploymentId)
-        && Objects.equals(apiVersion, that.apiVersion)
-        && Objects.equals(baseHost, that.baseHost)
-        && Objects.equals(path, that.path);
+        && Objects.equals(apiVersion, that.apiVersion);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(resourceName, deploymentId, apiVersion, baseHost, path,
-        useAzureApiKeyAuthentication, useAzureActiveDirectoryAuthentication);
+    return Objects.hash(resourceName, deploymentId, apiVersion, useAzureApiKeyAuthentication,
+        useAzureActiveDirectoryAuthentication);
   }
 }